+++ app/src/main/java/com/sipos/kebabsk/feature/transactions/domain/model/TransactionHistoryItem.kt
@@ -5,7 +5,7 @@ data class TransactionHistoryItem(
     val code: String,
     val time: String,
     val itemCount: Int,
-    val total: Double,
+    val total: Long,
     val status: String,
     val originalDate: String
 )
@@ -20,9 +20,9 @@ data class TransactionReceipt(
     val code: String,
     val createdAtLabel: String,
     val paymentMethod: String,
-    val totalAmount: Double,
-    val paidAmount: Double,
-    val changeAmount: Double,
+    val totalAmount: Long,
+    val paidAmount: Long,
+    val changeAmount: Long,
     val status: String,
     val items: List<TransactionReceiptItem>,
     val cashierName: String = "Kebab SK POS",
@@ -34,15 +34,15 @@ data class TransactionReceiptItem(
     val name: String,
     val variantName: String?,
     val qty: Int,
-    val price: Double,
-    val subtotal: Double
+    val price: Long,
+    val subtotal: Long
 )
 
 data class RevenueSummaryResult(
-    val totalRevenue: Double,
+    val totalRevenue: Long,
     val totalCount: Int,
     val transactionGrowthPercentage: Double?,
     val dominantItemName: String?,
     val revenueTargetPercentage: Double?,
-    val dailyTargetRevenue: Double?
+    val dailyTargetRevenue: Long?
 )
